renamed authors query; removed user details hook from deprecated
