--- src/domain/communication/communication/authorDetails.graphql
@@ -1,24 +1,0 @@
-query authorDetails($ids: [UUID!]!) {
-  users(IDs: $ids) {
-    id
-    firstName
-    lastName
-    isContactable
-    profile {
-      id
-      url
-      displayName
-      location {
-        id
-        country
-        city
-      }
-      visual(type: AVATAR) {
-        ...VisualUri
-      }
-      tagsets {
-        ...TagsetDetails
-      }
-    }
-  }
-}
