+++ src/domain/community/user/hooks/useAuthorsDetails.ts
@@ -1,6 +1,6 @@
 import { useCallback, useMemo } from 'react';
 import { AuthorModel } from '@/domain/community/user/models/AuthorModel';
-import { useAuthorDetailsQuery } from '@/core/apollo/generated/apollo-hooks';
+import { useUsersModelFullQuery } from '@/core/apollo/generated/apollo-hooks';
 import { uniq } from 'lodash';
 import { COUNTRIES_BY_CODE } from '@/domain/common/location/countries.constants';
 import { useAuthenticationContext } from '@/core/auth/authentication/hooks/useAuthenticationContext';
@@ -9,7 +9,7 @@ export const useAuthorsDetails = (authorIds: string[]) => {
   const uniqIds = uniq(authorIds).sort();
 
   const { isAuthenticated } = useAuthenticationContext();
-  const { data: authorData, loading } = useAuthorDetailsQuery({
+  const { data: authorData, loading } = useUsersModelFullQuery({
     variables: { ids: uniqIds },
     skip: uniqIds.length === 0 || !isAuthenticated,
   });
@@ -22,7 +22,7 @@ export const useAuthorsDetails = (authorIds: string[]) => {
           displayName: author.profile.displayName,
           firstName: author.firstName,
           lastName: author.lastName,
-          avatarUrl: author.profile.visual?.uri || '',
+          avatarUrl: author.profile.avatar?.uri || '',
           url: author.profile.url,
           tags: author.profile.tagsets?.flatMap(x => x.tags),
           city: author.profile.location?.city,
