+++ src/domain/community/user/hooks/useUserProvider.ts
@@ -1,18 +1,13 @@
 import { useUserModelFullQuery } from '@/core/apollo/generated/apollo-hooks';
 import { UserModel } from '@/domain/community/user/models/UserModel';
+
 /**
- * @deprecated Try to avoid this one, refactor to remove it
+ * Allows to retrieve a user, given its ID
  */
 export const useUserProvider = (userId: string | undefined) => {
   const { data, loading } = useUserModelFullQuery({
     variables: { userId: userId! },
     skip: !userId,
-    fetchPolicy: 'cache-and-network',
-    nextFetchPolicy: 'cache-first',
-    errorPolicy: 'all',
-    onError: () => {
-      // because reset store can crash - error needs to be consumed
-    },
   });
 
   const userModel: UserModel | undefined = data?.lookup?.user;
