+++ src/domain/community/user/hooks/userModelFull.graphql
