+++ src/domain/community/user/index.ts
@@ -1,3 +1,3 @@
-export * from '../../../_deprecated/useUserProvider';
+export * from './hooks/useUserProvider';
 export * from '../userCurrent/useCurrentUserContext';
 export * from '../userCurrent/CurrentUserProvider/usePlatformPrivilegeWrapper';
